# api/gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for ObservabilityScenarioResultOutcome.
const (
	Success ObservabilityScenarioResultOutcome = "success"
)

// Valid indicates whether the value is a known member of the ObservabilityScenarioResultOutcome enum.
func (e ObservabilityScenarioResultOutcome) Valid() bool {
	switch e {
	case Success:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	Error struct {
		// Code Stable application error code, independent of the HTTP status code.
		Code string `json:"code"`

		// Details Safe machine-readable context; keys use snake_case.
		Details   *map[string]interface{} `json:"details,omitempty"`
		Message   string                  `json:"message"`
		RequestId *string                 `json:"requestId,omitempty"`
	} `json:"error"`
}

// ObservabilityScenarioRequest defines model for ObservabilityScenarioRequest.
type ObservabilityScenarioRequest struct {
	// DelayMs Only accepted by slow-dependency; delay in the simulated dependency.
	DelayMs *int `json:"delayMs,omitempty"`

	// FailuresBeforeSuccess Only accepted by retry; failed attempts before eventual success.
	FailuresBeforeSuccess *int `json:"failuresBeforeSuccess,omitempty"`
}

// ObservabilityScenarioResult defines model for ObservabilityScenarioResult.
type ObservabilityScenarioResult struct {
	Attempts   int                                `json:"attempts"`
	DurationMs int64                              `json:"durationMs"`
	Outcome    ObservabilityScenarioResultOutcome `json:"outcome"`
	Scenario   string                             `json:"scenario"`

	// TraceId Omitted without a valid span context; not proof of export.
	TraceId *string `json:"traceId,omitempty"`
}

// ObservabilityScenarioResultOutcome defines model for ObservabilityScenarioResult.Outcome.
type ObservabilityScenarioResultOutcome string

// Task defines model for Task.
type Task struct {
	Completed bool      `json:"completed"`
	CreatedAt time.Time `json:"createdAt"`
	Id        string    `json:"id"`
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updatedAt"`
}

// CreateTaskJSONBody defines parameters for CreateTask.
type CreateTaskJSONBody struct {
	Title string `json:"title"`
}

// UpdateTaskJSONBody defines parameters for UpdateTask.
type UpdateTaskJSONBody struct {
	Completed bool `json:"completed"`
}

// RunObservabilityScenarioJSONRequestBody defines body for RunObservabilityScenario for application/json ContentType.
type RunObservabilityScenarioJSONRequestBody = ObservabilityScenarioRequest

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody CreateTaskJSONBody

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody UpdateTaskJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Run a bounded development-only observability demonstration
	// (POST /api/v1/observability/scenarios/{scenario}/run)
	RunObservabilityScenario(w http.ResponseWriter, r *http.Request, scenario string)
	// List tasks
	// (GET /api/v1/tasks)
	ListTasks(w http.ResponseWriter, r *http.Request)
	// Create a task
	// (POST /api/v1/tasks)
	CreateTask(w http.ResponseWriter, r *http.Request)
	// Update a task
	// (PATCH /api/v1/tasks/{id})
	UpdateTask(w http.ResponseWriter, r *http.Request, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// RunObservabilityScenario operation middleware
func (siw *ServerInterfaceWrapper) RunObservabilityScenario(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "scenario" -------------
	var scenario string

	err = runtime.BindStyledParameterWithOptions("simple", "scenario", r.PathValue("scenario"), &scenario, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "scenario", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RunObservabilityScenario(w, r, scenario)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTasks operation middleware
func (siw *ServerInterfaceWrapper) ListTasks(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTasks(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTask operation middleware
func (siw *ServerInterfaceWrapper) CreateTask(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTask(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdateTask operation middleware
func (siw *ServerInterfaceWrapper) UpdateTask(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateTask(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/observability/scenarios/{scenario}/run", wrapper.RunObservabilityScenario)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v1/tasks", wrapper.ListTasks)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/v1/tasks", wrapper.CreateTask)
	m.HandleFunc(http.MethodPatch+" "+options.BaseURL+"/api/v1/tasks/{id}", wrapper.UpdateTask)

	return m
}