# api/openapi.yaml
openapi: 3.0.3
info:
  title: GK API
  version: 0.1.0
servers:
  - url: /
paths:
  /api/v1/observability/scenarios/{scenario}/run:
    post:
      operationId: runObservabilityScenario
      summary: Run a bounded development-only observability demonstration
      description: Available only when APP_ENV=development. No external dependencies or database writes. At most four concurrent runs per process.
      parameters:
        - name: scenario
          in: path
          required: true
          schema:
            type: string
          description: One of baseline, slow-dependency, or retry. Unknown names return 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservabilityScenarioRequest"
            example: {}
      responses:
        "200":
          description: Completed scenario; intermediate retry failures may have occurred.
          headers:
            X-Trace-Id:
              description: Trace ID when a valid span context exists; does not guarantee export or sampling.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservabilityScenarioResult"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "408":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
        "500":
          $ref: "#/components/responses/Error"
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/tasks/{id}:
    patch:
      operationId: updateTask
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completed]
              properties:
                completed:
                  type: boolean
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
components:
  schemas:
    ObservabilityScenarioRequest:
      type: object
      additionalProperties: false
      properties:
        delayMs:
          type: integer
          minimum: 0
          maximum: 3000
          default: 1500
          description: Only accepted by slow-dependency; delay in the simulated dependency.
        failuresBeforeSuccess:
          type: integer
          minimum: 0
          maximum: 3
          default: 2
          description: Only accepted by retry; failed attempts before eventual success.
    ObservabilityScenarioResult:
      type: object
      required: [scenario, outcome, attempts, durationMs]
      properties:
        scenario:
          type: string
        outcome:
          type: string
          enum: [success]
        attempts:
          type: integer
        durationMs:
          type: integer
          format: int64
        traceId:
          type: string
          description: Omitted without a valid span context; not proof of export.
    Task:
      type: object
      required: [id, title, completed, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable application error code, independent of the HTTP status code.
              example: task_not_found
            details:
              type: object
              additionalProperties: true
              description: Safe machine-readable context; keys use snake_case.
            message:
              type: string
            requestId:
              type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"